requirements/base.txt
pydantic
pydantic-settings
python-dotenv
